requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
audioread==3.1.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comtypes==1.4.15
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
decorator==5.2.1
fastapi==0.129.0
fonttools==4.61.1
h11==0.16.0
idna==3.11
joblib==1.5.3
keyboard==0.13.5
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
matplotlib==3.10.8
msgpack==1.1.2
numba==0.63.1
numpy==2.3.5
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.1
platformdirs==4.7.0
pooch==1.9.0
PyAudio==0.2.14
pycodec2==4.1.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
pypiwin32==223
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pyttsx3==2.99
pywin32==311
qrcode==8.2
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
sounddevice==0.5.5
soundfile==0.13.1
soxr==1.0.0
SpeechRecognition==3.14.5
starlette==0.52.1
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0